Update mkimg.sh usage

+++ docker/setup.sh
@@ -6,7 +6,6 @@ deps=$1
 test "$deps" || deps=./bionic_deps.sh
 
 uid=$2
-test "$uid" || uid=1000
 
 source $deps
 
@@ -19,13 +18,16 @@ apt install -y --no-install-recommends software-properties-common
 add-apt-repository -y ppa:freecad-maintainers/freecad-daily
 apt update
 apt install -y --no-install-recommends $build_deps
-apt-get clean
-rm -rf /var/lib/apt/lists/* \
-   /usr/share/doc/* \
-   /usr/share/locale/* \
-   /usr/share/man/* \
-   /usr/share/info/*
 
-useradd -u $uid -ms /bin/bash freecad
-echo 'freecad:freecad' |chpasswd
+if test "$uid"; then
+    apt-get clean
+    rm -rf /var/lib/apt/lists/* \
+    /usr/share/doc/* \
+    /usr/share/locale/* \
+    /usr/share/man/* \
+    /usr/share/info/*
+
+    useradd -u $uid -ms /bin/bash freecad
+    echo 'freecad:freecad' |chpasswd
+fi
